Deprecate girder_client.HttpError in favor of requests.HTTPError
This deprecates a custom Girder class for HTTP exceptions with the one
provided by requests.

There was a subtle issue in our base sendRestRequest which would occur if a
request failed with an error code other than 404. In this scenario it
would try to return the JSON response which would work only some of
the time (independent of whether or not the request succeeded, since
error responses can return JSON). This led to confusing error messages
and other obscure issues like #1860. Now, requests either return a
response if successful or raise an error if there was a problem
(see requests.Response.raise_for_status).

Additionally the requests HTTPError provides callers with the
PreparedRequest and Response objects directly on it.